Former-commit-id: d94e99266d1e7bab0c5f5ad5888d66448185a0b6

+++ packages/pymedphys_dicom/tests/test_header_tweaks.py
@@ -35,7 +35,7 @@ import pydicom
 from pymedphys_dicom.dicom import (
     dicom_dataset_from_dict, adjust_machine_name, adjust_rel_elec_density,
     adjust_RED_by_structure_name, RED_adjustment_map_from_structure_names)
-from pymedphys_utilities.utilities import remove_file
+from pymedphys_dicom.utilities import remove_file
 
 
 HERE = os.path.dirname(__file__)
@@ -46,17 +46,21 @@ ADJUSTED_DICOM_FILENAME = os.path.join(
 
 
 def compare_dicom_cli(command, original, expected):
-    pydicom.write_file(ORIGINAL_DICOM_FILENAME, original)
+    # TODO: Extract CLI definition from subpackages and test separately
 
-    try:
-        subprocess.check_call(command)
-        cli_adjusted_ds = pydicom.read_file(
-            ADJUSTED_DICOM_FILENAME, force=True)
+    if 'SUBPACKAGE' not in os.environ:
 
-        assert str(cli_adjusted_ds) == str(expected)
-    finally:
-        remove_file(ORIGINAL_DICOM_FILENAME)
-        remove_file(ADJUSTED_DICOM_FILENAME)
+        pydicom.write_file(ORIGINAL_DICOM_FILENAME, original)
+
+        try:
+            subprocess.check_call(command)
+            cli_adjusted_ds = pydicom.read_file(
+                ADJUSTED_DICOM_FILENAME, force=True)
+
+            assert str(cli_adjusted_ds) == str(expected)
+        finally:
+            remove_file(ORIGINAL_DICOM_FILENAME)
+            remove_file(ADJUSTED_DICOM_FILENAME)
 
 
 def test_adjust_machine_name():
